feat: generic database interface
